requirements.txt
openai 
email-validator >= 2.0
fastapi==0.62.0
PyJWT==1.7.1
python-decouple==3.3
uvicorn==0.12.3
python-multipart
requests
chardet
aiofiles